refactor - improve method name
