+++ packages/helpers/src/find-duplicate-passcodes.ts
@@ -1,7 +1,13 @@
 import { ForceData } from '@serge/custom-types'
 import _ from 'lodash'
 
-const isUniquePasscode = (newData: ForceData, forces: ForceData[]): Array<{ forceName?:string, roleId?: string, roleName?: string }> => {
+/** check if any pass-codes in the new force being saved are duplicates
+ * of any in the existing data
+ * @param newData the force that has been modified
+ * @param forces the existing set of forces
+ * @returns list of roles with duplicate pass-codes
+ */
+const findDuplicatePasscodes = (newData: ForceData, forces: ForceData[]): Array<{ forceName?:string, roleId?: string, roleName?: string }> => {
   // use new force-data object, to check new values
   const forcesData = _.map(forces, force => force.uniqid === newData.uniqid ? newData : force)
 
@@ -32,4 +38,4 @@ const isUniquePasscode = (newData: ForceData, forces: ForceData[]): Array<{ forc
   return _.uniqBy(dupForceRoleNames, output => [output.forceName, output.roleName].join())
 }
 
-export default isUniquePasscode
+export default findDuplicatePasscodes
