+++ packages/helpers/src/index.ts
@@ -34,7 +34,7 @@ export {
 export { default as handleVisibilityAndConditionChanges } from './handle-visibility-condition-changes'
 export { default as hexNamed } from './hex-named'
 export { default as isChatChannel } from './is-chat-channel'
-export { default as isUniquePasscode } from './is-unique-passcode'
+export { default as findDuplicatePasscodes } from './find-duplicate-passcodes'
 export { default as getUniquePasscode } from './get-unique-passcode'
 export { default as isArrayOfObject } from './is-array-of-object'
 export { default as isBoolean } from './is-boolean'
