+++ packages/helpers/src/tests/find-duplicate-passcodes.spec.ts
@@ -1,6 +1,6 @@
 import { forces } from '@serge/mocks'
 import { ForceData } from '@serge/custom-types'
-import { isUniquePasscode } from '../'
+import { findDuplicatePasscodes } from '..'
 
 const allForces: ForceData[] = forces
 const selectedForce: ForceData = {
@@ -26,7 +26,7 @@ const selectedForce: ForceData = {
   ]
 }
 const selectedForceRevised: ForceData = {
-  ...forces[0],
+  ...forces[1],
   roles: [
     {
       name: 'CO',
@@ -57,7 +57,7 @@ const selectedForceRevised: ForceData = {
   ]
 }
 const selectedForceRevisedNew: ForceData = {
-  ...forces[0],
+  ...forces[2],
   roles: [
     {
       name: 'CO',
@@ -89,19 +89,19 @@ const selectedForceRevisedNew: ForceData = {
 }
 
 it('Check non-duplicate force role passcode', () => {
-  const res = isUniquePasscode(selectedForceRevisedNew, [selectedForceRevised])
+  const res = findDuplicatePasscodes(selectedForceRevisedNew, [selectedForceRevised])
   expect(res).toHaveLength(0)
 })
 
 it('Check multiple duplicate passcodes', () => {
-  const res = isUniquePasscode(selectedForceRevised, allForces)
+  const res = findDuplicatePasscodes(selectedForceRevised, allForces)
   expect(res).toHaveLength(4)
   expect(res[0].roleId).toEqual(res[1].roleId)
   expect(res[2].roleId).toEqual(res[3].roleId)
 })
 
 it('Check duplicate passcode between new force and rest of forces', () => {
-  const res = isUniquePasscode(selectedForce, allForces)
+  const res = findDuplicatePasscodes(selectedForce, allForces)
   expect(res).toHaveLength(2)
   expect(res[0].roleId).toEqual('rkrlss55f5e')
   expect(res[1].roleId).toEqual('rkrlss55f5e')
